优化

+++ object/funchelper.cpp
@@ -24,7 +24,7 @@ int8_t FuncHelper::byteToInt8(QByteArray by)
 int16_t FuncHelper::byteToInt16(QByteArray val,bool smallEnd)
 {
     int16_t tmp=0;
-    for(size_t i=0;i<2;i++){
+    for(int i=0;i<2;i++){
         int tmpI=smallEnd?1-i:i;
 
         tmp=(tmp <<8);
@@ -37,7 +37,7 @@ QByteArray FuncHelper::int16ToByte(int16_t val,bool smallEnd)
 {
     QByteArray by;
     by.resize(2);
-    for(size_t i=0;i<2;i++){
+    for(int i=0;i<2;i++){
         int tmpI=smallEnd?i:1-i;
         by[tmpI]=(val>>(i*8))&0xFF;
     }
@@ -47,7 +47,7 @@ QByteArray FuncHelper::int16ToByte(int16_t val,bool smallEnd)
 int32_t FuncHelper::byteToInt32(QByteArray val,bool smallEnd)
 {
     int32_t tmp=0;
-    for(size_t i=0;i<4;i++){
+    for(int i=0;i<4;i++){
         int tmpI=smallEnd?3-i:i;
 
         tmp=(tmp<<8);
@@ -60,7 +60,7 @@ QByteArray FuncHelper::int32ToByte(int32_t val,bool smallEnd)
 {
     QByteArray by;
     by.resize(4);
-    for(size_t i=0;i<4;i++){
+    for(int i=0;i<4;i++){
         int tmpI=smallEnd?i:3-i;
         by[tmpI]=(val>>(i*8))&0xFF;
     }
@@ -72,12 +72,12 @@ float FuncHelper::byteToFloat32(QByteArray val, bool smallEnd)
     float tmp=0;
     char* dest=(char*)(&tmp);
     if(isSmallEnd()){
-        for(size_t i=0;i<4;i++){
+        for(int i=0;i<4;i++){
             int tmpI=smallEnd?i:3-i;
             *(dest+i)=val[tmpI];
         }
     }else{
-        for(size_t i=0;i<4;i++){
+        for(int i=0;i<4;i++){
             int tmpI=smallEnd?3-i:i;
             *(dest+i)=val[tmpI];
         }
@@ -91,12 +91,12 @@ QByteArray FuncHelper::float32ToByte(float val, bool smallEnd)
     by.resize(4);
     char* dest=(char*)(&val);
     if(isSmallEnd()){
-        for(size_t i=0;i<4;i++){
+        for(int i=0;i<4;i++){
             int tmpI=smallEnd?i:3-i;
             by[tmpI]=*(dest+i);
         }
     }else{
-        for(size_t i=0;i<4;i++){
+        for(int i=0;i<4;i++){
             int tmpI=smallEnd?3-i:i;
             by[tmpI]=*(dest+i);
         }
@@ -109,12 +109,12 @@ double FuncHelper::byteToDouble64(QByteArray val, bool smallEnd)
     double tmp=0;
     char* dest=(char*)(&tmp);
     if(isSmallEnd()){
-        for(size_t i=0;i<8;i++){
+        for(int i=0;i<8;i++){
             int tmpI=smallEnd?i:7-i;
             *(dest+i)=val[tmpI];
         }
     }else{
-        for(size_t i=0;i<8;i++){
+        for(int i=0;i<8;i++){
             int tmpI=smallEnd?7-i:i;
             *(dest+i)=val[tmpI];
         }
@@ -128,12 +128,12 @@ QByteArray FuncHelper::double64ToByte(double val, bool smallEnd)
     by.resize(8);
     char* dest=(char*)(&val);
     if(isSmallEnd()){
-        for(size_t i=0;i<8;i++){
+        for(int i=0;i<8;i++){
             int tmpI=smallEnd?i:7-i;
             by[tmpI]=*(dest+i);
         }
     }else{
-        for(size_t i=0;i<8;i++){
+        for(int i=0;i<8;i++){
             int tmpI=smallEnd?7-i:i;
             by[tmpI]=*(dest+i);
         }
